use uintptr_t instead of unsigned int
looking in link.h on my ubuntu and f13 boxes both show in
/usr/include/link.h the type sig for la_objopen is

extern unsigned int la_objopen (struct link_map *__map, Lmid_t __lmid,
                uintptr_t *__cookie);

as opposed to unsigned int *. This has enabled compilation on my ubuntu
and fedora boxes

+++ audit.c
@@ -28,7 +28,7 @@ unsigned int la_version( unsigned int version )
 	return 1;
 }
 
-unsigned int la_objopen( struct link_map *lmp, Lmid_t lmid, unsigned int *cookie )
+unsigned int la_objopen( struct link_map *lmp, Lmid_t lmid, uintptr_t *cookie )
 {
 	static char *prefix = NULL;
